Move nested skills method to nested controller #95
